requirements.txt
langgraph==0.2.39
langchain
langchain-openai
langchain-community
openai==1.50.2
ollama==0.3.3
chromadb==0.5.5
fastapi==0.115.2
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
Pillow==10.4.0
pytesseract==0.3.13
opencv-python==4.10.0.84
python-barcode==0.15.1
pyzbar==0.1.9 # optional (requires libzbar)
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
streamlit==1.38.0